feat: support building npm/nodejs apps through the proxy
Fixes #26

One service handler is specifically for fetches of the tarball
artifacts -- these are verified against the policy server.

Another service handler acts as a pass-through proxy for all other
requests, per
https://github.com/npm/registry/blob/master/docs/REGISTRY-API.md

Noticed a bug in evaluate where jar files were the only thing ever
evaluated, so added a param for a required extension, which may be
different for every type of package.

Signed-off-by: Jim Crossley <jim@crossleys.org>